+++ docx4j-core/src/main/java/org/docx4j/model/properties/paragraph/KeepLines.java
@@ -1,0 +1,95 @@
+/*
+ *  Copyright 2009, Plutext Pty Ltd.
+ *   
+ *  This file is part of docx4j.
+
+    docx4j is licensed under the Apache License, Version 2.0 (the "License"); 
+    you may not use this file except in compliance with the License. 
+
+    You may obtain a copy of the License at 
+
+        http://www.apache.org/licenses/LICENSE-2.0 
+
+    Unless required by applicable law or agreed to in writing, software 
+    distributed under the License is distributed on an "AS IS" BASIS, 
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
+    See the License for the specific language governing permissions and 
+    limitations under the License.
+
+ */
+package org.docx4j.model.properties.paragraph;
+
+import org.docx4j.jaxb.Context;
+import org.docx4j.wml.BooleanDefaultTrue;
+import org.docx4j.wml.PPr;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.w3c.dom.Element;
+import org.w3c.dom.css.CSSValue;
+
+/**
+ * w:keepLines: all lines of the paragraph on one page.  XSL FO
+ * keep-together.within-page="always" (which also holds a paragraph whose
+ * w:br makes a nested block together, as Word does); CSS page-break-inside: avoid.
+ *
+ * @since 17.0.5
+ */
+public class KeepLines extends AbstractParagraphProperty {
+	
+	protected static Logger log = LoggerFactory.getLogger(KeepLines.class);	
+	
+	
+	public final static String CSS_NAME = "page-break-inside"; 
+	public final static String FO_NAME  = "keep-together.within-page"; 
+	
+	/**
+	 * @since 2.7.2
+	 */	
+	public String getCssName() {
+		return CSS_NAME;
+	}
+	
+	public KeepLines(BooleanDefaultTrue val) {
+		this.setObject(val);
+	}
+	
+	public KeepLines(CSSValue value) {	
+		
+		debug(CSS_NAME, value);
+		
+		if (value.getCssText().toLowerCase().equals("avoid")) {
+			this.setObject( Context.getWmlObjectFactory().createBooleanDefaultTrue()  );
+		} else {
+			BooleanDefaultTrue bdt = Context.getWmlObjectFactory().createBooleanDefaultTrue();
+			bdt.setVal(Boolean.FALSE);
+		}
+	}
+
+	@Override
+	public String getCssProperty() {
+		
+		if (((BooleanDefaultTrue)this.getObject()).isVal() ) {
+			return composeCss(CSS_NAME, "avoid");
+		} else {
+			return composeCss(CSS_NAME, "auto");
+		}
+	}
+
+
+	@Override
+	public void setXslFO(Element foElement) {
+
+		if (((BooleanDefaultTrue)this.getObject()).isVal() ) {
+			foElement.setAttribute(FO_NAME, "always" );
+		} else {
+			//
+		}
+		
+	}
+
+	@Override
+	public void set(PPr pPr) {
+		pPr.setKeepLines( (BooleanDefaultTrue)this.getObject() );
+	}
+	
+}